requirements.txt
streamlit==1.28.2
pydantic==2.5.0
pandas==2.0.3
pillow==10.1.0
matplotlib==3.7.2
numpy==1.24.3
easyocr==1.7.0
pdf2image==1.16.3
scikit-image==0.21.0
